backend/requirements.txt
# ── Web Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.20

# ── Database ──
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector>=0.3.0

# ── Redis ──
redis[hiredis]==5.2.1

# ── Configuration ──
pydantic-settings==2.7.1

# ── AI / Pipeline ──
pipecat-ai[silero,websocket,google]==0.0.105
openai>=1.50.0
deepgram-sdk>=4.0.0
elevenlabs>=1.0.0
google-genai>=1.0.0

# ── Telephony ──
twilio>=9.0.0

# ── Authentication ──
PyJWT>=2.8.0

# ── Utilities ──
httpx==0.28.1
python-dotenv==1.0.1
loguru>=0.7.0
tiktoken>=0.7.0

# ── Testing ──
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
aiosqlite>=0.20.0
fakeredis[lua]>=2.21.0
